+++ config/asset_compress.ini
@@ -49,4 +49,4 @@ files[] = sentences.edit_in_place.js
 files[] = sentences.change_language.js
 files[] = sentences.link.js
 files[] = links.add_and_delete.js
-files[] = collections.add_remove.js
+files[] = ratings.add_remove.js
