+++ config/auth_actions.php
@@ -7,10 +7,9 @@ $config = [
         'activities' => '*',
         'audio' => [ 'of', 'index' ],
         'autocompletions' => '*',
-        'collections' => [ 'of' ],
+        'ratings' => [ 'of' ],
         'contributions' => '*',
         'pages' => '*',
-        'collections' => [ 'of' ],
         'favorites' => [ 'of_user' ],
         's' => [ 's' ],
         'sentence_annotations' => [ 'last_modified' ],
@@ -75,7 +74,7 @@ $config = [
             'import' => [ User::ROLE_ADMIN ],
             'save_settings' => User::ROLE_CONTRIBUTOR_OR_HIGHER,
         ],
-        'collections'          => [ '*' => User::ROLE_CONTRIBUTOR_OR_HIGHER ],
+        'ratings'              => [ '*' => User::ROLE_CONTRIBUTOR_OR_HIGHER ],
         'exports'              => [ '*' => User::ROLE_CONTRIBUTOR_OR_HIGHER ],
         'favorites'            => [ '*' => User::ROLE_CONTRIBUTOR_OR_HIGHER ],
         'links'                => [ '*' => User::ROLE_ADV_CONTRIBUTOR_OR_HIGHER ],
