+++ src/Template/User/settings.ctp
@@ -44,7 +44,7 @@ $this->set('title_for_layout', $this->Pages->formatTitle(__('Settings')));
                 <h2><?php echo __('Options'); ?></h2>
             </div>
         </md-toolbar>
-        
+
         <?php echo $this->Form->create($userSettings, [
             'ng-cloak' => true,
             'url' => ['controller' => 'user', 'action' => 'save_settings']
@@ -256,7 +256,7 @@ $this->set('title_for_layout', $this->Pages->formatTitle(__('Settings')));
                     ng-init="collectionRatings = <?= $collectionRatings ?>"
                     class="md-primary">
                 </md-checkbox>
-                <p><?php echo __('Activate the feature to rate sentences and build your collection of sentences.') ?></p>
+                <p><?php echo __('Activate the feature to rate sentences.') ?></p>
                 <div ng-hide="true">
                 <?php
                     echo $this->Form->input(
